+++ apps/dashboard/components/hackerform/hackform-backnext-widget/hackform-backnext-widget.tsx
@@ -6,7 +6,7 @@ interface HackformBackNextWidgetProps {
   goNextQuestion: () => void;
 }
 
-const HackformBackNextWidget = ({
+export const HackformBackNextWidget = ({
   goNextQuestion,
   goPreviousQuestion,
 }: HackformBackNextWidgetProps) => {
