+++ apps/dashboard/components/hackerform/hackform-end/hackform-end.tsx
@@ -2,7 +2,7 @@ import { Colors2023 } from '@hibiscus/styles';
 import { FormMetadata } from '@hibiscus/types';
 import { H1, H3, Link } from '@hibiscus/ui';
 import { Button, GlowSpan } from '@hibiscus/ui-kit-2023';
-import { useHackform } from '../../hooks/use-hackform/use-hackform';
+import { useHackform } from '../../../hooks/use-hackform/use-hackform';
 import Image from 'next/image';
 import styled from 'styled-components';
