+++ apps/dashboard/components/hackerform/question-creator/question-creator.tsx
@@ -6,14 +6,14 @@ import {
   PageWrapper,
   QuestionWrapper,
   SmallText,
-} from './common-styled-components';
-import HackformQuestionHeader from './hackform-question-header';
-import HackformBackNextWidget from './hackform-backnext-widget';
+} from '../common-styled-components';
+import HackformQuestionHeader from '../hackform-question-header/hackform-question-header';
+import HackformBackNextWidget from '../hackform-backnext-widget/hackform-backnext-widget';
 import { FormQuestion, HackformError } from '@hibiscus/types';
 import { IoReturnDownBackOutline } from 'react-icons/io5';
 import styled from 'styled-components';
-import { GetInputResponseCb } from '../../common/types';
-import { useHackform } from '../../hooks/use-hackform/use-hackform';
+import { GetInputResponseCb } from '../../../common/types';
+import { useHackform } from '../../../hooks/use-hackform/use-hackform';
 
 interface QuestionCreatorProps {
   inputComponentChildren: React.ReactNode;
